pyproject.toml
[project]
name = "quiz-bots"
version = "0.1.0"
description = "Telegram & VK Quiz Bot"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "redis>=7.0.1",
    "vk-api>=11.10.0",
]
